Refactor TestGcdmNetcdfFileProblem into multiple test classes with more distinct names
