+++ cdm/gcdm/src/test/java/ucar/gcdm/TestGcdmFileFormats.java
@@ -1,0 +1,26 @@
+/*
+ * Copyright (c) 1998-2023 University Corporation for Atmospheric Research/Unidata
+ * See LICENSE for license information.
+ */
+package ucar.gcdm;
+
+import static ucar.gcdm.TestUtils.*;
+
+import java.lang.invoke.MethodHandles;
+import org.junit.Test;
+import org.junit.experimental.categories.Category;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import ucar.unidata.util.test.TestDir;
+import ucar.unidata.util.test.category.NeedsCdmUnitTest;
+
+public class TestGcdmFileFormats {
+  private static final Logger logger = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
+
+  @Test
+  @Category(NeedsCdmUnitTest.class)
+  public void testHdf4() throws Exception {
+    String localFilename = TestDir.cdmUnitTestDir + "formats/hdf4/MOD021KM.A2004328.1735.004.2004329164007.hdf";
+    compareVariables(localFilename, "MODIS_SWATH_Type_L1B/Data_Fields/EV_250_Aggr1km_RefSB");
+  }
+}
